+++ LookupAnything/release-notes.md
@@ -6,6 +6,7 @@
 * Added farm type description to player lookup.
 * Increased width of lookup UI.
 * Fixed planted coffee beans showing seed fields instead of crop fields.
+* Fixed seed growth time predictions not accounting for Agriculturist profession.
 
 ## 1.21.2
 Released 06 April 2019.